sr: zeroplus: Fix segfault, add debug output.
There are various ZEROPLUS models with different probe numbers. For now
hardcode to 16 (for the popular LAP-C(16032)). This will need to be
fixed in a dynamic way later.

This fixes a segfault due to only 16 probe-names being defined, but the
drivers returning 32 as probecount.

Also, add some additional debug output.